requirements.txt
appdirs==1.4.3
appnope==0.1.0
astroid==1.4.9
bleach==2.0.0
click==6.7
coverage==4.3.4
decorator==4.0.11
entrypoints==0.2.2
gitdb2==2.0.0
GitPython==2.1.3
html5lib==0.999999999
ipykernel==4.6.0
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.2.2
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
notebook==5.0.0
packaging==16.8
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
py==1.4.33
Pygments==2.2.0
pylint==1.6.5
pyparsing==2.2.0
pytest==3.0.7
pytest-cov==2.4.0
python-dateutil==2.6.0
pyzmq==16.0.2
qtconsole==4.3.0
simplegeneric==0.8.1
six==1.10.0
smmap2==2.0.1
terminado==0.6
testpath==0.3
tornado==4.4.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
wrapt==1.10.10
